3. Перегрузить метод max для трех чисел [#122644 #34974]

+++ src/main/java/ru/job4j/condition/Point.java
@@ -14,8 +14,8 @@ public class Point {
         this.y = second;
     }
 
-    public double distance(Point that) {
-        return sqrt(pow(this.x - that.x, 2) + pow(this.y - that.y, 2));
+    public int distance(Point that) {
+        return (int) sqrt(pow(this.x - that.x, 2) + pow(this.y - that.y, 2));
     }
 
     public static void main(String[] args) {
